+++ foghttp/_client/__init__.py
@@ -1,0 +1,3 @@
+"""Internal client implementation helpers."""
+
+__all__: tuple[str, ...] = ()
