+++ foghttp/types/__init__.py
@@ -1,0 +1,6 @@
+__all__ = (
+    "HttpVersion",
+    "HttpVersions",
+)
+
+from .http import HttpVersion, HttpVersions
