+++ foghttp/types/http.py
@@ -1,0 +1,10 @@
+__all__ = (
+    "HttpVersion",
+    "HttpVersions",
+)
+
+from typing import Literal, TypeAlias
+
+
+HttpVersion: TypeAlias = Literal["HTTP/1.1"]
+HttpVersions: TypeAlias = list[HttpVersion] | None
